requirements.txt
# Google Agent Development Kit
google-adk>=0.1.0

# A2A Protocol SDK for agent-to-agent communication
a2a-sdk[http-server]>=0.2.0
uvicorn>=0.30.0

# Langfuse for tracing and observability
langfuse>=2.0.0
openinference-instrumentation-google-adk>=0.1.0

# OpenTelemetry for comprehensive tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation>=0.40b0

# HTTP client for weather API
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
